+++ packages/crypto/src/validation/extensions/transactions/base.ts
@@ -64,7 +64,6 @@ export const base = joi =>
             .string()
             .alphanum()
             .required(),
-        signatures: joi.array(),
         secondSignature: joi.string().alphanum(),
         signSignature: joi.string().alphanum(), // TODO: remove in 2.1
         confirmations: joi // TODO: remove in 2.1